move endpoint call count processor to _trace, its only caller
